POC - do not merge!
Session window update hysteresis experiment:

Instead of sending FLOW performatives when the LWM is crossed, use
(hack) the LWM as a "maximum delta" to rate limit FLOW frame
generation.

When the difference between what the remote thinks is the window
capacity and the actual (local) capacity is >= LWM then send a FLOW
frame.